Issue #145  - Option to make the AL Object Browser the default 'editor' for .app files
